# static/swagger.yaml
openapi: "3.0.0"
info:
  title: Task Manager API
  version: "1.0.0"
  description: A simple API for managing tasks with user authentication.

paths:
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Registration successful
        "400":
          description: Username or email already exists

  /login:
    post:
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful with token
        "400":
          description: Invalid credentials

  /tasks:
    post:
      summary: Create a new task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                due_date:
                  type: string
      responses:
        "201":
          description: Task created

    get:
      summary: Retrieve all tasks for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of tasks for the user

  /tasks/{task_id}:
    put:
      summary: Update an existing task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
                priority:
                  type: string
                due_date:
                  type: string
      responses:
        "200":
          description: Task updated
        "404":
          description: Task not found

    delete:
      summary: Delete an existing task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task deleted
        "404":
          description: Task not found

  /logout:
    delete:
      summary: Logout the current user and revoke the token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Token successfully revoked

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT